test(sidekick/rust): split `annotate_test.go`
This file was definitely too large. Split the tests by "what annotation they
test". I am sure some functions are not quite perfect, but we can make
improvements over time.
